Rework ephemeral and binary heaps Implemented a sync.Cond for binary heap algo to save processor from spinning in the for loop and receiving nil Items until the Queue will be filled. Add num_pollers option to the configuration to specify number of pollers from the queue. Add Resume, ResumeAll, Stop, StopAll, PushBatch methods to the ephemeral. Remove map and use sync.Map in the ephemeral broker. Add protobuf schema.

+++ plugins/jobs/structs/job_options_test.go
@@ -79,10 +79,10 @@ func TestOptions_Merge(t *testing.T) {
 	})
 
 	assert.Equal(t, "pipeline", opts.Pipeline)
-	assert.Equal(t, 1, opts.Attempts)
-	assert.Equal(t, 2, opts.Delay)
-	assert.Equal(t, 1, opts.Timeout)
-	assert.Equal(t, 1, opts.RetryDelay)
+	assert.Equal(t, uint64(1), opts.Attempts)
+	assert.Equal(t, uint64(2), opts.Delay)
+	assert.Equal(t, uint64(1), opts.Timeout)
+	assert.Equal(t, uint64(1), opts.RetryDelay)
 }
 
 func TestOptions_MergeKeepOriginal(t *testing.T) {
@@ -103,8 +103,8 @@ func TestOptions_MergeKeepOriginal(t *testing.T) {
 	})
 
 	assert.Equal(t, "default", opts.Pipeline)
-	assert.Equal(t, 10, opts.Attempts)
-	assert.Equal(t, 10, opts.Delay)
-	assert.Equal(t, 10, opts.Timeout)
-	assert.Equal(t, 10, opts.RetryDelay)
+	assert.Equal(t, uint64(10), opts.Attempts)
+	assert.Equal(t, uint64(10), opts.Delay)
+	assert.Equal(t, uint64(10), opts.Timeout)
+	assert.Equal(t, uint64(10), opts.RetryDelay)
 }
